# *: simplify LIKE patterns with multiple consecutive wildcards