--- a/util/stringutil/string_util_test.go
+++ b/util/stringutil/string_util_test.go
@@ -144,8 +144,8 @@ func (s *testStringUtilSuite) TestCompileLike2Regexp(c *C) {
 		{`\_a`, `_a`},
 		{`\\_a`, `\.a`},
 		{`\a\b`, `\a\b`},
-		{`%%_`, `.*.`},
-		{`%_%_aA`, ".*..*.aA"},
+		{`%%_`, `..*`},
+		{`%_%_aA`, "...*aA"},
 	}
 	for _, v := range tbl {
 		result := CompileLike2Regexp(v.pattern)
@@ -181,20 +181,50 @@ func (s *testStringUtilSuite) TestIsExactMatch(c *C) {
 	}
 }
 
-func BenchmarkMatchSpecial(b *testing.B) {
-	var (
-		pattern = `a%a%a%a%a%a%a%a%b`
-		target  = `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa`
-		escape  = byte('\\')
-	)
+func BenchmarkDoMatch(b *testing.B) {
+	escape := byte('\\')
+	tbl := []struct {
+		pattern string
+		target  string
+	}{
+		{`a%_%_%_%_b`, `aababab`},
+		{`%_%_a%_%_b`, `bbbaaabb`},
+		{`a%_%_a%_%_b`, `aaaabbbbbbaaaaaaaaabbbbb`},
+	}
 
-	patChars, patTypes := CompilePattern(pattern, escape)
+	for _, v := range tbl {
+		b.Run(v.pattern, func(b *testing.B) {
+			patChars, patTypes := CompilePattern(v.pattern, escape)
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				match := DoMatch(v.target, patChars, patTypes)
+				if !match {
+					b.Fatal("Match expected.")
+				}
+			}
+		})
+	}
+}
 
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		match := DoMatch(target, patChars, patTypes)
-		if match {
-			b.Fatal("Unmatch expected.")
-		}
+func BenchmarkDoMatchNegative(b *testing.B) {
+	escape := byte('\\')
+	tbl := []struct {
+		pattern string
+		target  string
+	}{
+		{`a%a%a%a%a%a%a%a%b`, `aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa`},
+	}
+
+	for _, v := range tbl {
+		b.Run(v.pattern, func(b *testing.B) {
+			patChars, patTypes := CompilePattern(v.pattern, escape)
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				match := DoMatch(v.target, patChars, patTypes)
+				if match {
+					b.Fatal("Unmatch expected.")
+				}
+			}
+		})
 	}
 }