### What problem does this PR solve?

Problem Summary: `%% => %` and `%_ => _%` are valid rewrite rules that improve pattern matching performance.

This is a follow up to https://github.com/pingcap/tidb/pull/17418#issuecomment-634553964 (cc @wjhuang2016).

### What is changed and how it works?

What's Changed:

* LIKE pattern compile now applies the rules `%% => %` and `%_ => _%`.

### Related changes
Preceding: https://github.com/pingcap/tidb/pull/17418

### Check List

Tests

- Unit test

Moderate performance improvement for the affected cases:

```
name                                 old time/op  new time/op  delta
DoMatch/a%_%_%_%_b-8                 31.0ns ± 0%  20.2ns ± 0%   ~     (p=1.000 n=1+1)
DoMatch/%_%_a%_%_b-8                 37.1ns ± 0%  26.2ns ± 0%   ~     (p=1.000 n=1+1)
DoMatch/a%_%_a%_%_b-8                 142ns ± 0%    94ns ± 0%   ~     (p=1.000 n=1+1)
DoMatchNegative/a%a%a%a%a%a%a%a%b-8   193ns ± 0%   171ns ± 0%   ~     (p=1.000 n=1+1)
```

### Release note
* Simplify LIKE patterns with multiple consecutive wildcards.